+++ CHANGELOG.md
@@ -11,10 +11,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Documentation of specifications
 - Laravel Integration
 - Travis-CI Integration
+- Ability to Post[Feature]
 
 ###Fixed
 - Fixed RegistrationTest
 - Typo in setup.md
+- Fixed Traffic Tab width too large(in dashboard)
 
 ###Changed
 - Changed graphs to Chart.js